CAMEL-13629: Renamed tests to follow *Test pattern
