+++ components/camel-crypto/src/test/java/org/apache/camel/component/crypto/SpringSignatureTest.java
@@ -30,14 +30,14 @@ import org.apache.camel.spring.SpringCamelContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
 
-public class SpringSignatureTest extends SignatureTests {
+public class SpringSignatureTest extends SignatureTest {
 
     private static KeyPair rsaPair;
 
     protected CamelContext createCamelContext() throws Exception {
         rsaPair = getKeyPair("RSA");
         return SpringCamelContext.springCamelContext(
-                new ClassPathXmlApplicationContext("org/apache/camel/component/crypto/SpringSignatureTests.xml"), true);
+                new ClassPathXmlApplicationContext("org/apache/camel/component/crypto/SpringSignatureTest.xml"), true);
     }
 
     public static KeyStore keystore() throws Exception {
@@ -45,7 +45,7 @@ public class SpringSignatureTest extends SignatureTests {
     }
 
     public static PrivateKey privateKeyFromKeystore() throws Exception {
-        return new SignatureTests().getKeyFromKeystore();
+        return new SignatureTest().getKeyFromKeystore();
     }
 
     public static Certificate certificateFromKeystore() throws Exception {
