+++ components/camel-crypto/src/test/resources/org/apache/camel/component/crypto/SpringSignatureTest.xml
