+++ components/camel-netty4-http/src/test/java/org/apache/camel/component/netty4/http/NettyHttpBindingUseAbsolutePathTest.java
@@ -24,7 +24,7 @@ import org.apache.camel.component.http4.HttpMethods;
 import org.junit.Test;
 
 
-public class NettyHttpBindingUseAbsolutePath extends BaseNettyTest {
+public class NettyHttpBindingUseAbsolutePathTest extends BaseNettyTest {
 
     @Test
     public void testSendToNettyWithPath() throws Exception {
