+++ components/camel-soroush/src/test/java/org/apache/camel/component/soroushbot/component/ProducerDownloadFileTest.java
@@ -17,12 +17,14 @@
 package org.apache.camel.component.soroushbot.component;
 
 import java.io.ByteArrayInputStream;
-import java.util.Map;
+import java.io.IOException;
+import java.io.InputStream;
 
 import org.apache.camel.EndpointInject;
 import org.apache.camel.RoutesBuilder;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.apache.camel.component.soroushbot.IOUtils;
 import org.apache.camel.component.soroushbot.models.MinorType;
 import org.apache.camel.component.soroushbot.models.SoroushAction;
 import org.apache.camel.component.soroushbot.models.SoroushMessage;
@@ -30,12 +32,15 @@ import org.apache.camel.component.soroushbot.support.SoroushBotTestSupport;
 import org.apache.camel.component.soroushbot.support.SoroushBotWS;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
-public class ProducerAutoUploadFileIsFalse extends SoroushBotTestSupport {
+@Ignore("CAMEL-13629 failing test")
+public class ProducerDownloadFileTest extends SoroushBotTestSupport {
 
     @EndpointInject("direct:soroush")
     org.apache.camel.Endpoint endpoint;
+
 
     @Override
     @Before
@@ -49,23 +54,20 @@ public class ProducerAutoUploadFileIsFalse extends SoroushBotTestSupport {
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                from("direct:soroush").to("soroush://" + SoroushAction.sendMessage + "/token?autoUploadFile=false")
+                from("direct:soroush").to("soroush://" + SoroushAction.sendMessage + "/token")
                         .process(exchange -> {
                             SoroushMessage body = exchange.getIn().getBody(SoroushMessage.class);
-                            if (body.getFileUrl() != null) {
-                                throw new AssertionError("file url is null");
-                            }
-                            if (body.getThumbnailUrl() != null) {
-                                throw new AssertionError("thumb url is null");
-                            }
+                            body.setFile((InputStream) null);
+                            body.setThumbnail((InputStream) null);
                         })
+                        .to("soroush://" + SoroushAction.downloadFile + "/token")
                         .to("mock:soroush");
             }
         };
     }
 
     @Test
-    public void autoUploadTest() throws Exception {
+    public void checkDownloadFile() throws InterruptedException, IOException {
         SoroushMessage body = new SoroushMessage();
         body.setType(MinorType.TEXT);
         body.setFrom("b1");
@@ -78,11 +80,9 @@ public class ProducerAutoUploadFileIsFalse extends SoroushBotTestSupport {
         MockEndpoint mockEndpoint = getMockEndpoint("mock:soroush");
         mockEndpoint.setExpectedMessageCount(1);
         mockEndpoint.assertIsSatisfied();
-        Assert.assertEquals("message sent successfully", SoroushBotWS.getReceivedMessages().get(0), body);
         SoroushMessage mockedMessage = mockEndpoint.getExchanges().get(0).getIn().getBody(SoroushMessage.class);
-        Map<String, String> fileIdToContent = SoroushBotWS.getFileIdToContent();
-        Assert.assertEquals("file uploaded successfully", fileIdToContent.size(), 0);
-        Assert.assertEquals(mockedMessage.getFileUrl(), null);
-        Assert.assertEquals(mockedMessage.getThumbnailUrl(), null);
+        Assert.assertEquals("download file successfully", new String(IOUtils.readFully(mockedMessage.getFile(), 1000, false)), fileContent);
+        Assert.assertEquals("download thumbnail successfully", new String(IOUtils.readFully(mockedMessage.getThumbnail(), 1000, false)), thumbContent);
     }
+
 }
